requirements.txt
flask
tensorflow
numpy
opencv-python-headless
pytesseract
Pillow
groq
gunicorn
python-dotenv
